requirements.txt
FastAPI
uvicorn
pydantic
sqlalchemy
python-dotenv
google-generativeai